requirements.txt
snowflake-connector-python==3.0.0
pandas==2.0.3
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2
